code: add duplicate object feature * This code was entirely written by Codex with the model GPT-5 High. Prompts:
1.----------------------------------------------------------------------------
So there is this thing called `Storage2`, it have named objects and unnamed objects that are `inline` somewhere else. Everything is stored by id.

I want to create "duplicate" button for the Storage2, so that it creates duplicate of element that it stores. But I want to make it in such a way that all the inner contents is also duplicated. Take for example `enum Object`, it have `MatrixId` inside. If it's a `inline` matrix, then when duplicating this object, I want inner matrix is also to be duplicated in its corresponding `Storage2<Matrix>`, and have a new id. But `Matrix` itself have a lot of other things that can be `inline`. And so on recursively.

Can you analyze the code and suggest a plan how to do that? Maybe we need to have a new type of trait, that allows for such modifications? Don't write the code yet, only analyze the code and suggest a plan that you think will work.

2.----------------------------------------------------------------------------
okay, let's implement this plan, also please add this duplicate button to the storage2::egui element egui, near the `Delete` button

3.----------------------------------------------------------------------------
please remove default implementation of StorageElem2::duplicate_inline and implement this for all other types